Updated paths due to code move
