+++ server/src/routes/termdb.chat.ts
@@ -6,8 +6,8 @@ import { classifyQuery } from './chat/classify1.ts'
 import { classifyPlotType } from './chat/plot.ts'
 import { classifyNotPlot } from './chat/classify2.ts'
 import { inferScaffold } from './chat/scaffold.ts'
-import serverconfig from '../src/serverconfig.js'
-import { getDsAllowedTermTypes } from './termdb.config.ts'
+import serverconfig from '../serverconfig.js'
+import { getDsAllowedTermTypes } from '../../routes/termdb.config.ts'
 import { phrase2entity } from './chat/phrase2entity.ts'
 import { inferTermObjFromEntity } from './chat/entity2termObj.ts'
 import { resolveToTwTvs } from './chat/entity2twTvs.ts'
